+++ R/benchmark_enhancer_predictor.R
@@ -116,11 +116,11 @@
 #' @param database epiRomics class database
 #'   containing all data initially loaded
 #' @param histone name or vector of histone
-#'   mark(s), must match name in epiROmics_dB@meta,
-#'   default set to h3k4me1
+#'   mark(s), must match a \code{name} entry in
+#'   \code{\link{meta}(database)}, default set to h3k4me1
 #' @param curated_database database to test
-#'   histone marks against, must match name in
-#'   epiROmics_dB@meta default set to fantom
+#'   histone marks against, must match a \code{name} entry in
+#'   \code{\link{meta}(database)}, default set to fantom
 #' @return Variable of class dataframe further exploring top histone marks that
 #'   may determine enhancer regions
 #' @export